move tsoBatchController to tso_batch_controller.go to improve code readability of client.go
